fix: separate partial comment type definition for zod2openapi unknown error
