+++ packages/openapi/src/comment/get.ts
@@ -2,10 +2,23 @@ import type { RouteConfig } from '@asteasolutions/zod-to-openapi';
 import { axios } from '../axios';
 import { registerRoute, urlBuilder } from '../utils';
 import { z } from '../zod';
-import { commentSchema } from './types';
-import type { ICommentVo } from './types';
+import { commentReactionSchema } from './reaction';
+import { commentContentSchema } from './types';
 
 export const GET_COMMENT_DETAIL = '/comment/{tableId}/{recordId}/{commentId}';
+
+export const commentSchema = z.object({
+  id: z.string(),
+  content: commentContentSchema,
+  createdBy: z.string(),
+  reaction: commentReactionSchema.optional().nullable(),
+  createdTime: z.date(),
+  lastModifiedTime: z.date(),
+  quoteId: z.string().optional(),
+  deletedTime: z.date().optional(),
+});
+
+export type ICommentVo = z.infer<typeof commentSchema>;
 
 export const GetCommentDetailRoute: RouteConfig = registerRoute({
   method: 'get',
@@ -22,7 +35,7 @@ export const GetCommentDetailRoute: RouteConfig = registerRoute({
       description: "Returns the record's comment detail",
       content: {
         'application/json': {
-          schema: z.array(commentSchema),
+          schema: commentSchema.array(),
         },
       },
     },
