+++ packages/openapi/src/comment/index.ts
@@ -1,9 +1,9 @@
 export * from './get-list';
 export * from './types';
 export * from './create';
-export * from './delete';
 export * from './update';
 export * from './get';
+export * from './delete';
 export * from './reaction';
 export * from './notify';
 export * from './get-attachment-url';
