+++ packages/openapi/src/comment/reaction/create-reaction.ts
@@ -2,10 +2,30 @@ import type { RouteConfig } from '@asteasolutions/zod-to-openapi';
 import { axios } from '../../axios';
 import { registerRoute, urlBuilder } from '../../utils';
 import { z } from '../../zod';
-import type { IUpdateCommentReactionRo } from '../types';
-import { updateCommentReactionRoSchema } from '../types';
+import type { EmojiSymbol } from './constant';
+import { SUPPORT_EMOJIS } from './constant';
 
 export const CREATE_COMMENT_REACTION = '/comment/{tableId}/{recordId}/{commentId}/reaction';
+
+export const commentReactionSymbolSchema = z
+  .string()
+  .emoji()
+  .refine((value) => {
+    return SUPPORT_EMOJIS.includes(value as EmojiSymbol);
+  });
+
+export const commentReactionSchema = z
+  .object({
+    reaction: commentReactionSymbolSchema,
+    user: z.array(z.string()),
+  })
+  .array();
+
+export const updateCommentReactionRoSchema = z.object({
+  reaction: commentReactionSymbolSchema,
+});
+
+export type IUpdateCommentReactionRo = z.infer<typeof updateCommentReactionRoSchema>;
 
 export const CreateCommentReactionRoute: RouteConfig = registerRoute({
   method: 'post',
